move comment tests to own file, do some cleanup
